- made the QueryView code more modular.
